Update heading context step to support GOV.UK Design System
As part of migrating our frontends to use GOV.UK Design System styles
and components we are replacing instances of our heading context style
with the GOV.UK Design System caption style.

We want to be able to have the test working on either version, so this
commit updates the step that looks for heading contexts to also look for
captions.