Update 'Log out' steps
In alphagov/digitalmarketplace-user-frontend#155 we decided that we
wanted the logout button in the service header to be a link, to make it
easier for the element to be accessible.

Now we are replacing govuk_template and the old header with `dmHeader`
from alphagov/digitalmarketplce-govuk-frontend#25, which does use a link
for the logout button.

This has the side effect of breaking tests that expect 'Log out' to be a
button or input rather than an anchor; this commit updates the
functional tests to be indifferent to the HTML element used.