Fix steps that were looking for a h2 on a page
We had an issue where the presence of the cookie banner was causing an
ambiguous match for scenarios looking for a h2 on the page, because the
cookie banner includes a h2.

This commit changes the step that looks for things on the page to look
within the main element only, which excludes the cookie banner.