Add partial release support alphagov/digitalmarketplace-buyer-frontend#1030
This can be reverted once alphagov/digitalmarketplace-buyer-frontend#1030
has been released.

We want the tests to pass on both preview and staging with slightly
different content for the checkbox 'I understand that I cannot edit my
search results( again)? after I export my results'. Normally to do this
we would have two versions of the test, one with the new content and one
with the old content, and add appropriate tags for each stage.

Unfortunately this particular step is part of another multi-step step
definition, which makes that process a bit harder. Instead I've added a
new step that targets the specific checkbox using the form input name
(which hasn't changed).

It's a bit of a hack, but hopefully it won't be in our code for long!