Use Bash as Makefile shell
Currently we seem to use the default - sh. This causes problems with the default `local.sh` since `BASH_SOURCE` is not valid sh.

Switch to force the use of bash to ensure that `BASH_SOURCE` works.

Note that `local.sh` does have a bash shebang. However, this is useless because the shebang is not evaluated when the script is sourced.