Complete copied service
For DOS 5, we introduced a bug that broke services that were copied from previous frameworks. This bug only showed up if you tried to complete your copied service for submission, so did not show up in our automated tests. So complete the copied service and mark it as complete during the test.