Change step to create a saved search to use api
Use the api instead of substeps for `When I have created and saved a
search`. This has the advantage of working even if the user journey to
save a search changes.

Also, avoid creating a saved search if one with the same name already
exists, so we don't end up with users with loads of saved searches all
called 'my cloud project'.