Change date format for brief response availability
In DOS4 and previous frameworks, the availability field was a free text entry.
In DOS5 is is now a Date question and has day, month and year input separately.

When availability is loaded from a saved, but not sumbitted brief response
it must be in the YYYY-MM-DD format so it can be parsed and filled into
the correct form fields. Change the format on the fixture, and update
DOS4 tests to look for the new value.