Make it simpler and more robust to get a supplier
Previously, we would get the supplier by id/name/registered name and it assume it had the right data, or create it with the right data.

This caused problems because it was possible for the system to be in a state where the supplier the test got had the wrong data for the test. Also, no tests get supplier by ID.

So simplify the helper by removing the ability to get by ID.

Also make it more robust by always doing an update to ensure that the supplier we got/created has the data we expect.