Update is_validation_header? method for govuk-frontend
We recently changed the supplier frontend to use govuk-frontend error
summary instead of frontend toolkit validation banners[[1]].

This commit changes the functional tests to also look for an error
summary when checking if a validation banner is present. This wasn't a
problem before because we only use this machinery for testing the
framework application journeys.

[1]: https://github.com/alphagov/digitalmarketplace-supplier-frontend/pull/1340