Make it simpler and more robust to get a user
Currently, we fail the test if a supplier exists but has the wrong properties. This was a deliberate choice: ce1a702a5b10428a727d24a596969e7d32d18cfa.

> I realize that in cases we find a property mismatch, we could actually go ahead and update the found user to match the requested properties. I've held back on doing this for now though - it has the potential to create a lot of churn on our db and allow us to introduce more race problems for parallel tests.

This has come back to bite us. We've just seen a case where a user got into the wrong state, so the tests would fail for the rest of time without manual intervention. Also, I disagree with the original author about database churn and race conditions - the only time we would do a database update is when the tests would otherwise fail, so there's not much churn. And I don't think we run our tests in parallel - and these tests couldn't be run in parallel anyway due to operating on the same users.

Thus, I think we can safely switch to new logic where we update the user if we discover that they exist but don't match what we expect. This should mean that this test becomes a bit more robust.

Note that this could cause problems if we ever used this helper method against a user in production. However, that's not something we do in any of our smoke or smoulder tests, so I think we're safe.

Follows similar logic to 9abdd8b, which did the same thing for suppliers.