Fix accessible autocomplete step
Fix the step for entering text into an accessible autocomplete field.

This stopped working when we bumped accessible-autocomplete to 1.0.1 in
the admin frontend[[1]]. It appears that using the `set` method doesn't
work with the JavaScript for this version of the accessible
autocomplete, as the typed option was not being focused.

This commit fixes it by using the `send_key` method to emulate a user
typing; this works with both the newer and older version of
accessible-autocomplete.

[1]: https://ci.marketplace.team/job/functional-tests-preview/23280/functional_20test_20report/