Add assertion that there is only one copied service
For the "Supplier copies a service from a previous framework" feature
occasionally I get a failure where a service that has already been
copied is copied again. I think this is because I am testing with G12
open and G11 live, and as G12 has already been made live something funky
is going on. Unfortunately I don't have time to debug it fully.

This commit adds a RSpec expectation that a draft service should have
only one copy per framework, so when this failure occurs it is clear to
the developer. Maybe one day this error will start someone else on the
journey to the solution.