Add command to check feature syntax and run in CI
We just had a bunch of test failures because a PR was merged with a syntax error in a feature. This was not noticed before merging because the CI only linted the Ruby code.

Add a make command to parse (but not execute) all the features. This should catch future syntax errors. Run this command as part of pre-merge CI.