Fix 'regex on page' step
At the end of https://github.com/alphagov/digitalmarketplace-functional-tests/pull/892 I made a small change to tidy it up, but neglected to re-test. Inevitably, that small change broke the whole thing.

It turns out that trying to re-use the 'text on page' step doesn't work - the regex gets converted back into text. So fix it by copying the necessary parts across. I've tested that this works this time.