Complete new question on staging
In https://github.com/alphagov/digitalmarketplace-functional-tests/pull/894
we duplicated this feature to account for the new question that's on
preview but not on staging.

Once https://github.com/alphagov/digitalmarketplace-briefs-frontend/pull/447
has been released to staging, we no longer need different versions of
the feature.

Remove the skip tags and update the feature so it always answers the
new question.