Allow Selenium to upload files
Selenium can't interact with non-visible elements, which causes problems
with file uploads as described in this [Stack Overflow answer](https://stackoverflow.com/questions/53395806/capybara-with-selenium-and-remote-chrome-how-to-attach-file-for-upload)

We can trick Selenium into doing what we want by temporarily making the
file input visible before trying to add the file. Thankfully this is a
common enough problem that Capybara has a method for it.