Get Design System 3 page count
The previous xpath worked for the old toolkit, but not for the new Design System 3. We've now switched over all search pages to Design System 3, so don't need to be back-compatible.

The previous behaviour was concealing a bug because the xpath never found anything, so it assumed that there was only one page of results.