Add a smoulder test to test AWS connection
Some apps have AWS credentials to allow them to upload/download from private S3 buckets. It would be useful to have a smoulder test that verifies that the AWS credentials are working correctly. Admin download of supplier details is the only endpoint readily available at all points in the framework lifecycle, so we can use that for our testing.